refactor(swimlane): split swimlane components
Move bloated ConfigChangesSwimlane presentation pieces into focused components under the Swimlane directory. Keep the main swimlane component responsible for state, bucketing, grouping, and orchestration while preserving existing behavior.
