+++ src/components/Configs/Changes/Swimlane/EmptySwimlaneState.tsx
@@ -1,0 +1,7 @@
+export function EmptySwimlaneState() {
+  return (
+    <div className="flex h-full items-center justify-center text-gray-400">
+      No changes to display
+    </div>
+  );
+}
