+++ src/components/Configs/Changes/Swimlane/FlexLabel.tsx
@@ -1,0 +1,5 @@
+export function FlexLabel({ text }: { text: string }) {
+  return (
+    <span className="whitespace-nowrap text-xs text-gray-400">{text}</span>
+  );
+}
