+++ src/components/Configs/Changes/Swimlane/Legend.tsx
