+++ src/components/Configs/Changes/Swimlane/LoadingSwimlaneState.tsx
@@ -1,0 +1,8 @@
+export function LoadingSwimlaneState() {
+  return (
+    <div className="flex h-full items-center justify-center gap-2 text-sm text-gray-400">
+      <div className="h-4 w-4 animate-spin rounded-full border-2 border-gray-300 border-t-gray-600" />
+      Loading changes...
+    </div>
+  );
+}
